+++ lib/constants.dart
@@ -24,6 +24,7 @@ const kMaxShownPresets = 10;
 const kMaxNSIPresets = 3;
 const kFollowLinks = false;
 const kUploadOnClose = false;
+const kMicromappingTapZoom = 19.0;
 
 const kDefaultPresets = [
   'shop/convenience', 'shop/clothes',